liquidity: Move functions from liquidity formula test to helper.
This is so we can use these functions in other places as well, e.g in liquidity
conversion rate test.
